requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3

# LLM / AI
openai==1.3.5
anthropic==0.7.1
tiktoken==0.5.1
requests==2.31.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Security & Validation
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Rate Limiting
slowapi==0.1.9

# Logging & Monitoring
python-json-logger==2.0.7

# Data Export
pandas==2.1.3

# Testing (for future use)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1

# Development
python-dotenv==1.0.0
